requirements.txt
gradio
requests
gradio
requests
pandas
langchain-core
langchain-google-genai
langchain-google-community
langgraph

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
et_xmlfile==2.0.0
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-modelarmor==0.3.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-experimental==0.4.0
langchain-google-community==3.0.0
langchain-google-genai==3.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
llvmlite==0.45.1
MarkupSafe==3.0.3
marshmallow==3.26.1
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx
numba==0.62.1
numpy
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
openpyxl==3.1.5
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xlrd==2.0.2
xxhash==3.6.0
yarl==1.22.0
youtube-transcript-api==1.2.3
zstandard==0.25.0